Applied new structure
