+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/country_list/view_holder/chart_helper/DayValueFormatter.kt
@@ -1,0 +1,13 @@
+package com.moshenskyi.feature_covid_data.country_list.view_holder.chart_helper
+
+import com.github.mikephil.charting.components.AxisBase
+import com.github.mikephil.charting.data.LineData
+import com.github.mikephil.charting.formatter.ValueFormatter
+
+class DayValueFormatter(private val dataSet: LineData) : ValueFormatter() {
+
+	override fun getAxisLabel(value: Float, axis: AxisBase?): String {
+		val entryForIndex = dataSet.getDataSetByIndex(0).getEntryForIndex(value.toInt())
+		return entryForIndex.data.toString()
+	}
+}
