--- covid-data/src/main/java/com/moshenskyi/feature_covid_data/data/model/general/CountryRecords.kt
@@ -1,3 +1,0 @@
-package com.moshenskyi.feature_covid_data.data.model.general
-
-class CountryRecords : ArrayList<CountryRecordsItem>()
