--- covid-data/src/main/java/com/moshenskyi/feature_covid_data/data/model/vaccine/VaccineCapacityInfo.kt
@@ -1,3 +1,0 @@
-package com.moshenskyi.feature_covid_data.data.model.vaccine
-
-class VaccineCapacityInfo : ArrayList<VaccineCapacityInfoItem>()
