--- covid-data/src/main/java/com/moshenskyi/feature_covid_data/data/model/vaccine/VaccineCapacityInfoItem.kt
@@ -1,6 +1,0 @@
-package com.moshenskyi.feature_covid_data.data.model.vaccine
-
-data class VaccineCapacityInfoItem(
-    val country: String,
-    val timeline: Map<String, Long>
-)
