--- covid-data/src/main/java/com/moshenskyi/feature_covid_data/domain/model/CovidInfoEntity.kt
@@ -1,11 +1,0 @@
-package com.moshenskyi.feature_covid_data.domain.model
-
-import com.moshenskyi.feature_covid_data.data.model.vaccine.VaccineCapacityInfoItem
-
-data class CovidInfoEntity(
-    val testCount: Int,
-    val country: String,
-    val population: Int,
-    val vaccineCapacityInfoItem: VaccineCapacityInfoItem?,
-    var expanded: Boolean
-)
