+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/model/general/CountryRecords.kt
@@ -1,0 +1,3 @@
+package com.moshenskyi.feature_covid_data.model.general
+
+class CountryRecords : ArrayList<CountryRecordsItem>()
