+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/model/general/CountryRecordsItem.kt
@@ -1,0 +1,28 @@
+package com.moshenskyi.feature_covid_data.model.general
+
+data class CountryRecordsItem(
+	val active: Int,
+	val activePerOneMillion: Double,
+	val cases: Int,
+	val casesPerOneMillion: Int,
+	val continent: String,
+	val country: String,
+	val countryInfo: CountryInfo,
+	val critical: Int,
+	val criticalPerOneMillion: Double,
+	val deaths: Int,
+	val deathsPerOneMillion: Double,
+	val oneCasePerPeople: Int,
+	val oneDeathPerPeople: Int,
+	val oneTestPerPeople: Int,
+	val population: Int,
+	val recovered: Int,
+	val recoveredPerOneMillion: Double,
+	val tests: Int,
+	val testsPerOneMillion: Int,
+	val todayCases: Int,
+	val todayDeaths: Int,
+	val todayRecovered: Int,
+	val undefined: Double,
+	val updated: Long
+)
