+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/model/vaccine/VaccineCapacityInfo.kt
@@ -1,0 +1,3 @@
+package com.moshenskyi.feature_covid_data.model.vaccine
+
+class VaccineCapacityInfo : ArrayList<VaccineCapacityInfoItem>()
