+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/model/vaccine/VaccineCapacityInfoItem.kt
@@ -1,0 +1,6 @@
+package com.moshenskyi.feature_covid_data.model.vaccine
+
+data class VaccineCapacityInfoItem(
+    val country: String,
+    val timeline: Map<String, Long>
+)
