+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/network/CovidRepositoryImpl.kt
@@ -1,0 +1,19 @@
+package com.moshenskyi.feature_covid_data.network
+
+import com.moshenskyi.feature_covid_data.domain_entity.CovidRepository
+import com.moshenskyi.feature_covid_data.model.general.CountryRecordsItem
+import com.moshenskyi.feature_covid_data.model.vaccine.VaccineCapacityInfoItem
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOf
+
+class CovidRepositoryImpl : CovidRepository {
+	private val api = RetrofitBuilder.api
+
+	override suspend fun getCountriesInfo(): Flow<List<CountryRecordsItem>> {
+		return flowOf(api.getCountriesInfo())
+	}
+
+	override suspend fun getVaccinationCapacity(): Flow<List<VaccineCapacityInfoItem>> {
+		return flowOf(api.getVaccinationCapacity())
+	}
+}
